+++ components/PageHead.js
@@ -5,6 +5,7 @@ export default function PageHead(props) {
     title = "LinkFree - connect to your audience with a single link",
     description = "Showcase the content you create and your projects in one place. Make it easier for people to find, follow and subscribe.",
     ogTitle = "LinkFree - connect to your audience with a single link",
+    ogDescription = "Showcase the content you create and your projects in one place. Make it easier for people to find, follow and subscribe.",
     ogUrl = "http://linkfree.eddiehub.io",
     ogImage = "/logo512.png",
     ogType = "image/png",
@@ -19,6 +20,7 @@ export default function PageHead(props) {
       <link rel="icon" href="/favicon.ico" />
 
       <meta property="og:title" content={ogTitle} />
+      <meta property="og:description" content={ogDescription} />
       <meta property="og:type" content={ogType} />
       <meta property="og:url" content={ogUrl} />
       <meta property="og:image" content={ogImage} />