+++ pages/events.js
@@ -5,36 +5,14 @@ import { TbCoin, TbCoinOff } from "react-icons/tb";
 
 import { getEvents } from "./api/events";
 
-import logger from "@config/logger";
 import EventCard from "@components/event/EventCard";
 import Page from "@components/Page";
 import { EventTabs } from "@components/event/EventTabs";
 import PageHead from "@components/PageHead";
 import Badge from "@components/Badge";
 
-export async function getServerSideProps(context) {
+export async function getServerSideProps() {
   let events = await getEvents();
-
-  // remove any invalid events
-  events = events.filter((event) => {
-    const dateTimeStyle = {
-      dateStyle: "full",
-      timeStyle: "long",
-    };
-    try {
-      new Intl.DateTimeFormat("en-GB", dateTimeStyle).format(
-        new Date(event.date.start)
-      );
-      new Intl.DateTimeFormat("en-GB", dateTimeStyle).format(
-        new Date(event.date.end)
-      );
-
-      return true;
-    } catch (e) {
-      logger.error(e, `ERROR event date for: "${event.name}"`);
-      return false;
-    }
-  });
 
   return {
     props: { events },
@@ -46,9 +24,7 @@ export default function Events({ events }) {
     all: events,
     virtual: events.filter((event) => event.isVirtual === true),
     inPerson: events.filter((event) => event.isInPerson === true),
-    cfpOpen: events.filter((event) =>
-      event.date.cfpClose ? new Date(event.date.cfpClose) > new Date() : false
-    ),
+    cfpOpen: events.filter((event) => event.date.cfpOpen === true),
     free: events.filter((event) => event.price?.startingFrom === 0),
     paid: events.filter((event) => event.price?.startingFrom > 0),
   };
@@ -129,7 +105,7 @@ export default function Events({ events }) {
           {categorizedEvents[eventType]?.map((event) => (
             <EventCard
               event={event}
-              username={event.username}
+              usernames={event.usernames}
               key={`${event.name} ${event.username}`}
             />
           ))}