+++ pages/index.js
@@ -1,5 +1,7 @@
 import Image from "next/image";
 import { IconContext } from "react-icons";
+import Script from "next/script";
+
 import {
   MdOutlinePlayArrow,
   MdHelpOutline,
@@ -10,28 +12,19 @@ import {
 } from "react-icons/md";
 import { FaMedal } from "react-icons/fa";
 
-import singleUser from "@config/user.json";
 import { getTodayStats } from "./api/statistics/today";
 import { getTotalStats } from "./api/statistics/totals";
 import Link from "@components/Link";
 import PageHead from "@components/PageHead";
 import BasicCards from "@components/statistics/BasicCards";
-import Button from "@components/Button";
 import Testimonials from "@components/Testimonials";
 import GitHubAccelerator from "@components/GitHubAccelerator";
 import Alert from "@components/Alert";
 import config from "@config/app.json";
+import CallToAction from "@components/CallToAction";
 
 export async function getStaticProps() {
   const pageConfig = config.isr.homepage; // Fetch the specific configuration for this page
-  if (singleUser.username) {
-    return {
-      redirect: {
-        destination: `/${singleUser.username}`,
-        permanent: true,
-      },
-    };
-  }
 
   const { stats: totalStats } = await getTotalStats();
   const { stats: todayStats } = await getTodayStats();
@@ -249,12 +242,12 @@ export default function Home({ total, today }) {
         />
       </div>
 
-      <div className="bg-white dark:bg-primary-high">
+      <div className="bg-primary-low dark:bg-primary-high">
         <div className="mx-auto max-w-7xl py-16 px-4 sm:px-6 lg:px-8">
           <div className="overflow-hidden rounded-lg bg-secondary-high shadow-xl lg:grid lg:grid-cols-2 lg:gap-4">
             <div className="px-6 pt-10 pb-12 sm:px-16 sm:pt-16 lg:py-16 lg:pr-0 xl:py-20 xl:px-20">
               <div className="lg:self-center">
-                <h1 className="text-3xl font-bold tracking-tight text-white sm:text-4xl">
+                <h1 className="text-3xl font-bold tracking-tight text-primary-low sm:text-4xl">
                   <span className="block">Connect to your audience</span>
                   <span className="block">with a single link</span>
                 </h1>
@@ -265,50 +258,35 @@ export default function Home({ total, today }) {
                 </p>
               </div>
             </div>
-            <div className="aspect-w-5 aspect-h-3 -mt-6 md:aspect-w-2 md:aspect-h-1">
-              <Image
-                className="translate-x-6 translate-y-6 transform rounded-md object-cover object-left-top sm:translate-x-16 h-auto w-auto"
-                src="https://user-images.githubusercontent.com/624760/230707268-1f8f1487-6524-4c89-aae2-ab45f0e17f39.png"
-                priority
-                alt="App screenshot"
-                width={500}
-                height={500}
-              />
+            <div className="aspect-w-16 aspect-h-9">
+              <div
+                className="kartra_video_containergbHEDtAnMwlF js_kartra_trackable_object"
+                data-kt-type="video"
+                data-kt-value="gbHEDtAnMwlF"
+                data-kt-owner="nkmvj7Xr"
+              ></div>
+              <Script src="https://app.kartra.com/video/gbHEDtAnMwlF"></Script>
             </div>
           </div>
         </div>
       </div>
 
-      <div className="bg-primary-low dark:bg-primary-high">
-        <div className="mx-auto max-w-7xl py-12 px-4 sm:px-6 lg:flex lg:items-center lg:justify-between lg:py-16 lg:px-8">
-          <h2 className="text-3xl font-bold tracking-tight text-primary-high dark:text-primary-low sm:text-4xl">
-            <span className="block">Ready to dive in?</span>
-            <span className="block text-secondary-high dark:text-secondary-low">
-              Add your free Profile today!
-            </span>
-          </h2>
-          <div className="mt-8 flex lg:mt-0 lg:flex-shrink-0">
-            <div className="inline-flex rounded-md shadow">
-              <Button
-                text="Get started"
-                href="/docs/quickstart"
-                primary={true}
-              />
-            </div>
-            <div className="ml-3 inline-flex rounded-md shadow ">
-              <Button text="Example" href="/eddiejaoude" />
-            </div>
-          </div>
-        </div>
-      </div>
+      <CallToAction
+        title="Ready to dive in?"
+        description="Add your free Profile today!"
+        button1Link="/docs/quickstart"
+        button1Text="Get started"
+        button2Link="/eddiejaoude"
+        button2Text="Example"
+      />
 
       <div className="bg-secondary-high">
         <div className="mx-auto max-w-2xl py-12 px-4 sm:px-6 sm:py-12 lg:max-w-7xl lg:px-8">
           <div className="mx-auto max-w-3xl text-center">
-            <h2 className="text-3xl font-bold tracking-tight text-white sm:text-4xl">
+            <h2 className="text-3xl font-bold tracking-tight text-primary-low sm:text-4xl">
               Features
             </h2>
-            <p className="mt-4 text-white">
+            <p className="mt-4 text-primary-low text-xl">
               It is not just links... Take a look at the Features you can add to
               customize your LinkFree Profile.
             </p>
@@ -328,10 +306,10 @@ export default function Home({ total, today }) {
                     "mt-6 lg:mt-0 lg:row-start-1 lg:col-span-5 xl:col-span-4"
                   )}
                 >
-                  <h3 className="text-lg font-medium text-white">
+                  <h3 className="text-lg sm:text-2xl font-medium text-primary-low">
                     {feature.name}
                   </h3>
-                  <p className="mt-2 text-sm text-white">
+                  <p className="mt-2 text-sm sm:text-lg text-primary-low">
                     {feature.description}
                   </p>
                 </div>
@@ -343,13 +321,12 @@ export default function Home({ total, today }) {
                     "flex-auto lg:row-start-1 lg:col-span-7 xl:col-span-8"
                   )}
                 >
-                  <div className="aspect-w-5 aspect-h-2 overflow-hidden rounded-lg bg-primary-low">
+                  <div className="aspect-w-5 aspect-h-2 overflow-hidden rounded-lg bg-primary-low relative">
                     <Image
                       src={feature.imageSrc}
                       alt={feature.imageAlt}
-                      className="object-cover object-center"
-                      width={1250}
-                      height={840}
+                      className="object-contain object-center"
+                      fill={true}
                     />
                   </div>
                 </div>
@@ -359,7 +336,7 @@ export default function Home({ total, today }) {
         </div>
       </div>
 
-      <div className="relative bg-white dark:bg-primary-high py-24 sm:py-32 lg:py-40">
+      <div className="relative bg-primary-low dark:bg-primary-high py-8 sm:py-12 lg:py-24">
         <div className="mx-auto max-w-md px-6 text-center sm:max-w-3xl lg:max-w-7xl lg:px-8">
           <h2 className="font-semibold text-secondary-high dark:text-secondary-low text-3xl">
             Getting Started
@@ -385,12 +362,12 @@ export default function Home({ total, today }) {
                         <div>
                           <span className="inline-flex items-center justify-center rounded-xl bg-secondary-high p-3 shadow-lg">
                             <feature.icon
-                              className="h-8 w-8 text-white "
+                              className="h-8 w-8 text-primary-low"
                               aria-hidden="true"
                             />
                           </span>
                         </div>
-                        <h3 className="mt-8 text-lg font-semibold leading-8 tracking-tight group-hover:underline group-hover:text-secondary-medium dark:text-white">
+                        <h3 className="mt-8 text-lg font-semibold leading-8 tracking-tight group-hover:underline group-hover:text-secondary-medium dark:text-primary-low">
                           {feature.name}
                         </h3>
                         <p className="mt-5 text-base leading-7 text-primary-high dark:text-primary-low-medium">
@@ -408,16 +385,24 @@ export default function Home({ total, today }) {
 
       <Testimonials data={testimonials} />
 
+      <CallToAction
+        title="Subscribe to our newsletter to learn more"
+        description="Do not miss out!"
+        button1Link="/newsletter"
+        button1Text="Sign up"
+      />
+
       <GitHubAccelerator />
 
       <Link
         href="https://github.com/EddieHubCommunity/LinkFree/discussions"
         rel="noopener noreferrer"
         target="_blank"
+        className="fixed bottom-5 right-5 rounded-full focus-visible:outline focus-visible:outline-2 focus-visible:outline-offset-2 focus-visible:outline-secondary-medium"
       >
-        <div className="fixed bottom-5 right-5 px-4 py-2 bg-secondary-medium text-white flex items-center gap-1 rounded-full hover:drop-shadow-lg hover:bg-secondary-high-high">
+        <div className="px-4 py-2 bg-secondary-medium text-primary-low flex items-center gap-1 rounded-full hover:drop-shadow-lg hover:bg-secondary-high-high">
           <IconContext.Provider
-            value={{ color: "white", style: { verticalAlign: "middle" } }}
+            value={{ color: "primary-low", style: { verticalAlign: "middle" } }}
           >
             <MdHelpOutline />
           </IconContext.Provider>