+++ CHANGELOG.md
@@ -1,9 +1,35 @@
-## 1.133.8 (2023-05-29)
+# [1.146.0](https://github.com/EddieHubCommunity/LinkFree/compare/v1.145.5...v1.146.0) (2023-06-29)
+
+
+### Features
+
+* manage profile with forms ([#6998](https://github.com/EddieHubCommunity/LinkFree/issues/6998)) ([82bb960](https://github.com/EddieHubCommunity/LinkFree/commit/82bb9603874db8f7b11322a853ac78c5100a7cea)), closes [#7699](https://github.com/EddieHubCommunity/LinkFree/issues/7699)
+
+
+
+## [1.145.5](https://github.com/EddieHubCommunity/LinkFree/compare/v1.145.4...v1.145.5) (2023-06-28)
+
+
+
+## [1.145.4](https://github.com/EddieHubCommunity/LinkFree/compare/v1.145.3...v1.145.4) (2023-06-28)
 
 
 ### Bug Fixes
 
-* parent div to header ([#7155](https://github.com/EddieHubCommunity/LinkFree/issues/7155)) ([fdaf3e6](https://github.com/EddieHubCommunity/LinkFree/commit/fdaf3e64c44596d09b84482a074b39396c8856e8))
+* storybook user components ([#7802](https://github.com/EddieHubCommunity/LinkFree/issues/7802)) ([a63405a](https://github.com/EddieHubCommunity/LinkFree/commit/a63405a8d154d00f6f77bddf55cccb1162c8d6d1))
+
+
+
+## [1.145.3](https://github.com/EddieHubCommunity/LinkFree/compare/v1.145.2...v1.145.3) (2023-06-27)
+
+
+### Bug Fixes
+
+* bg color event drop down filter ([#7803](https://github.com/EddieHubCommunity/LinkFree/issues/7803)) ([e2fe1f1](https://github.com/EddieHubCommunity/LinkFree/commit/e2fe1f149be39ee16a5e2734fa28118352b3448d))
+
+
+
+## [1.145.2](https://github.com/EddieHubCommunity/LinkFree/compare/v1.145.1...v1.145.2) (2023-06-26)